Move jQuery handlers into one place
