+++ src/ggrc/assets/javascripts/common-handlers.js
@@ -6,6 +6,55 @@
 import Spinner from 'spin.js';
 
 var $body = $('body');
+var $window = $(window);
+
+// We remove loading class
+$window.on('load', function () {
+  $('html').removeClass('no-js');
+});
+
+$body.on("click", ".lhn-no-init", function() {
+  $(this).removeClass('lhn-no-init');
+  import(/* webpackChunkName: "lhn" */'./controllers/lhn_controllers')
+    .then(function () {
+      $("#lhn").cms_controllers_lhn();
+      $(document.body).ggrc_controllers_recently_viewed();
+    });
+});
+
+$body.on("click", "a[data-toggle=unmap]", function(ev) {
+  var $el = $(this)
+  ;
+  //  Prevent toggling `openclose` state in trees
+  ev.stopPropagation();
+  $el.fadeTo('fast', 0.25);
+  $el.children(".result").each(function(i, result_el) {
+    var $result_el = $(result_el)
+      , result = $result_el.data('result')
+      , mappings = result && result.get_mappings()
+      , i
+    ;
+
+    function notify(instance){
+      $(document.body).trigger(
+        "ajax:flash"
+        , {"success" : "Unmap successful."}
+      );
+    }
+
+    can.each(mappings, function(mapping) {
+      mapping.refresh().done(function() {
+        if (mapping instanceof CMS.Models.Control) {
+          mapping.removeAttr('directive');
+          mapping.save().then(notify);
+        }
+        else {
+          mapping.destroy().then(notify);
+        }
+      });
+    });
+  });
+});
 
 // Initialize delegated event handlers
 jQuery(function ($) {
