+++ src/ggrc/assets/javascripts/entrypoints/dashboard/bootstrap.js
@@ -294,9 +294,4 @@
       model_class: notesModel
     });
   });
-
-  // We remove loading class
-  $(window).on('load', function () {
-    $('html').removeClass('no-js');
-  });
 })(window.can.$, window.can, window.CMS, window.GGRC);
